cleanig CommBlock: converted face refresh counter to use charm Loop class; precompute face counter in count_refresh_() instead of recomputing; simplified is_on_boundary(), update_boundary_(); fixed warnings in ItBlock and mesh_Hierarchy.cpp; BUG FIX--added missing counter update in mesh_RefineSlope (added before somewhere?); reverted Monitor to output time
--HG--
branch : add-refine

+++ src/Cello/mesh_Hierarchy.cpp
@@ -44,7 +44,7 @@ Hierarchy::~Hierarchy() throw()
 
 # else
 
-  for (int i=0; i<block_.size(); i++) {
+  for (size_t i=0; i<block_.size(); i++) {
     delete block_[i];
     block_[i] = 0;
   }
